Examples: Rename GroundProjectedEnv to GroundProjectedSkybox.
